Refactor Phase 4: extract input/events into viewer/input.js
Move all user interaction (mouse, touch, keyboard, wheel, overlay
buttons, overlay menu) from starfield.js into a dedicated input
module. Move applyDragRotation to camera.js where it belongs with
the other projection math. starfield.js drops from ~1085 to ~866 lines.
